pyproject.toml
[project]
name = "snaq-verify"
version = "0.1.0"
description = "Agentic nutrition verification for SNAQ food items"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "SNAQ Task" }]

dependencies = [
    "pydantic>=2.10",
    "pydantic-ai-slim[openai]>=0.0.30",
    "httpx>=0.28",
    "tenacity>=9.0",
    "typer>=0.15",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "respx>=0.21",
    "ruff>=0.8",
]

[project.scripts]
snaq-verify = "snaq_verify.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/snaq_verify"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
